Updated the URL to http://localhost:8080/docs/
The URL http://localhost:8080/documentation/ throws a 404 that reads:
404 PAGE NOT FOUND
The web page or file you were trying to reach cannot be found. It may have been moved to a different location or removed because the information it contained is no longer valid.

Common errors that can result in broken links include:

- Truncated URL's. Some email applications may break a URL due to line breaks. You may need to cut and paste a URL from your email application into your browser's location field.

- Case Sensitivity. URL's are case sensitive. "File.html" and "file.html" are not considered the same file by the server. Please verify that the URL is entered correctly as printed.

Please email your feedback/comments.