Fix the error when Okta wrongly rejects validation
Problem: when making a request to create a claim with filtering by groups if you follow the guide you will get a parsing validation error. 
You must change `groupFilterType` in the request to `group_filter_type`. It works then. Okta folks need to have a uniform guidance on this.

Example:
```
const formData = {
    "name": "groupName",
    "status": "ACTIVE",
    "claimType": "RESOURCE",
    "valueType": "GROUPS",
    "group_filter_type": "EQUALS",
    "value": groupName
  }
```