Updating index.md to refer to system API
I've added the warning from the event types page to the event hooks page. I was originally looking to use the event hooks to capture information using a webhook pattern. The changes to event types and system log API seem important enough to call out at the beginning of the Event Hooks documentation not a level in when looking at the event types