Deprecate the existing SAML guide
The existing SAML guide content has been replaced with the combined SAML/OIDC guide created under the OIN umbrella. This change removes the older guide and updates all redirects and direct links to that guide.

The staging server will be used to test the redirects are functional before merging to the live code base.

Resolves OKTA-279606