Clarify the difference between "App" and "Integration" in SCIM Guide
This update changes most of the files in the SCIM guide, concept and reference topics to clarify when we mean an "integration" (an Okta thing) and and "application" (a customer's cloud application). There was a fair amount of ambiguity in the use of these terms throughout our SCIM documentation, so this set of changes aims to correct that.

Resolves [OKTA-297173](https://oktainc.atlassian.net/browse/OKTA-297173)