Clarifies on the deletion action in SCIM applications
In several places we discuss the DELETE operation for SCIM provisioning, but there is not a clear outline of what happens when the deletion of the user resource occurs on the Okta integration side, or on the customer's SCIM application end.

A single set of rules has been laid out in the concepts topic, and all references to deletions in the SCIM guide and reference topics have been updated to point to that set of rules.

Resolves [OKTA-179948](https://oktainc.atlassian.net/browse/OKTA-179948)