Update index.md
The instructions are a little unclear how to use group claims for non Okta mastered groups, or if it can be done at all. I linked to an external knowledge article which details the process. I assume the information of this article may need to be incorporated into the docs if external links are discouraged.