Fix omission re modifying base properties
When we made the `format` field of the `login` property modifiable (to support non-email logins), we documented it under "User Profile base subschema" but missed another place that links to there: the description of the "Update User Profile Schema property" command.  I felt that the list of allowed modifications there was getting unwieldy, so I rearranged the sentence to make it more concise.