Removed non-existent anchor link
This link leads to an anchor that does not exist. I pulled up google cache and found it used to just say, look below, so I don't think it was ever meant to be a link. Please let me know if I'm off here. (reported by Tim Lopez in #web)