Update index.md
I don't see this point mentioned anywhere in the docs and have seen a few cases now where customers don't know the idp parameter nullifies the login_hint parameter. There's currently an Okta idea to have this changed but not sure if it has traction: https://ideas.okta.com/app/#/case/110841". I leave it up to you whether it has a place in this doc. Thanks!