Not supported: Explicit references to apps and custom claims
There may be a more efficient way to call attention to this besides repeating the same note twice - but I think it should be stated that the explicit reference to apps will not work when creating custom claims for OIDC/OAuth. This recently tripped up a customer of mine. If you ping me on Okta slack I can share a convo I had about this as well internally. Thanks!