Update configmid.md
Update documentation on configuring Okta.AspNetCore package to start off with adding the using statement as the first step.  This will provide a better user experience to people going through the steps as it will enable IntelliSense if supported by their IDE.

Also removed the $ symbol in the statements setting the OktaDomain property of the OktaMvcOptions.