Update index.md
Removed the line implying you no longer need to configure a Trusted Origin if you add a "Login redirect URI" to an application in the admin console. This was only true in the Developer console and now you MUST add the domain to Trusted Origins for the widget to work no matter what.