Update index.md
The current documentation makes it seem like a user should build the URL with all the query params themselves. It is not clear that there behind the scene events happening which generate and store PKCE codes. I have added documentation to suggest using the SDK as to help make it clear to other developers.