Update index.md (#4100)
As your company grows, the number of user accounts and provisioned software applications increases, but it does not increase exponentially. While the word _exponentially_ can be used in a non-literal sense to mean _rapidly_ or  _greatly_, it should be avoided in technical writing where the reader (who's probably an engineer) is likely to assume the word should be interpreted in the mathematical sense.