General updates for mobile items
Updated the mobile versions for all three articels by either removing extraneous SIW content or adding a note that SIW is unsupported. This is done by either adding more stack snippets or modifying the content of existing snippets or the main article (index.md).

This commit also contains the first drafts for "Set up your Okta org" and "Run the sample apps." The final article requires a lot of change to the code portions.