Update index.md
I started this test by researching Okta, reading through the Style Guide, reading the article and researching what I did not understand, and visualized how the content could be improved with structural changes. 

Research included reading documents from the Internet, watching You Tube videos, and reviewing the type of documents 
you have in the help center to compare.

The topic is a concept, but focused on new features. That indicated to me that this was not a typical concept overview article. It read a lot like a What's New (a type of release note where new features are announced). Although it had more content than a release note. I decided to structure the article so that the user would be able to understand the conceptual material about a new product they might want to upgrade to. When I write What's New articles in my current position, they are meant to be a short summary of the new feature, benefits, and then link out to additional information. The What's New is eventually retired and when retired, the content lives on in the Help Center in the conceptual article.

The following are highlights of my editing excercise in the order that I made the updates:

- Title: Removed the "overview" from the title. The concept title should be able to stand on its own and users should be able to recognize that the title defines what they will find in the article. Adding overview to title is redundant.

-Rewrote the short description to combine the first few paragraphs.

-Always looking to write in a minimalistic approach. Less words, more common words, etc. Removing redundancies, jargon, and other componund sentences is better for user experience and for translation to global docs.

- Repeated use of Okta - I cleaned up some of usage of the term. Okta is in the title, so we can assume that the whole article is about Okta. I kept the first occurrence of the full name of the product and some usage of Okta after that, but overall removed from all other instances of the Identity Engine name.

-Jargon - Removed several instances of jargon, including the word "pipeline", the phrase "step up"

- Overall changed the article to read in active voice. Tried to keep in active voice as much as possible.

- Removed words like valuable, recommended, as these words might come back to bite. All clients might not think it is valuable or they can come back and say "you said this was valuable, or you said this was the recommended way" etc. I get mixed reviews about telling the user that one method is recommended over the other. In the case of the deployment models. I removed that the Okta-hosted Sign-In Widget was recommended and kept strong words to indicate they should considered using this method.

-Because I titled the article as a new features article, I removed the Identity Engine new features section heading. 

- I continued on to structure the content in three main  sections: Features, Deployment models, Documentation support, and Upgrade to Identity Engine. I also added links to provide an easier flow through the documentation.

- Within the Features section, I create a section for each feature with links.

- I removed parenthesis or brackets from around the example given for the sign-in page. I couldn't find a style guideline around parenthesis in your guide, but I follow MS Word Style Guide and they recommend to not use parenthesis, just use commas. I applied this in other areas also.

-I searched for common nouns in the Okta Help Center to make sure I was following the right protocol if I did not see in your guide. Examples...made sure that Sign-In Widget was written appropriately.


Overall, this was a fun, but challenging test. I use an editor on a daily basis and had to do a quick refresher about HTML and markup. The knowledge came back quickly, but took me a little longer to complete the excercise because I wanted to do the restructure and add the links.

Thanks again for the opportuntiy.